+++ arc/context.py
@@ -1,14 +1,16 @@
 from __future__ import annotations
 import contextlib
+import os
 import typing as t
 
 from arc import errors, logging, typing as at
 from arc.callback import CallbackStack
 from arc.color import colorize, fg
 from arc.config import config
-from arc.types.params import special
 from arc import _command
+from arc.params import special
 from arc.types.state import State
+from arc.prompt import Prompt
 
 
 if t.TYPE_CHECKING:
@@ -24,6 +26,8 @@ default_sug: at.Suggestions = {
     "suggest_arguments": True,
     "suggest_commands": True,
 }
+
+default_prompt = Prompt(" ")
 
 
 @special(default=object())
@@ -44,6 +48,10 @@ class Context:
             `CLI` root. If a command is executing standalone, this will always be empty
         execute_callbacks (bool): whether or not to execute command callbacks
             when executing the command
+        env_prefix: (str, optional): A prefix to use when selecting values from environmental
+            variables. Will be combined with the name specified for a parameter.
+        prompt: (arc.prompt.Prompt, optional): A prompt object will be used when prompting
+            for parameter values.
     """
 
     # Each time a command is invoked,
@@ -52,6 +60,7 @@ class Context:
     # context will be popped off the stack
     __stack: list[Context] = []
     config = config
+    _meta: dict[str, t.Any] = {}
 
     def __init__(
         self,
@@ -61,12 +70,18 @@ class Context:
         parent: Context = None,
         suggestions: at.Suggestions = None,
         execute_callbacks: bool = True,
+        env_prefix: str = "",
+        prompt: Prompt = None,
     ):
         self.command = command
         self.fullname = fullname
         self.command_chain = command_chain or []
         self.parent = parent
         self.execute_callbacks = execute_callbacks
+        if "prompt" not in self.meta:
+            self.meta["prompt"] = prompt or default_prompt
+        if "env_prefix" not in self.meta:
+            self.meta["env_prefix"] = env_prefix
 
         if suggestions is not None:
             self.suggestions = default_sug | suggestions
@@ -108,6 +123,21 @@ class Context:
             curr = curr.parent
 
         return curr
+
+    @property
+    def meta(self):
+        """Meta is an object that is
+        shared globally by all `Context` instances
+        """
+        return self._meta
+
+    @property
+    def prompt(self) -> Prompt:
+        return self.meta["prompt"]
+
+    @property
+    def env_prefix(self) -> str:
+        return self.meta["env_prefix"]
 
     def create_state(self) -> State:
         state: State = State()
@@ -186,6 +216,9 @@ class Context:
     def exit(self, code: int = 0) -> t.NoReturn:
         """Exits the app with code `code`"""
         raise errors.Exit(code)
+
+    def getenv(self, name: str, default):
+        return os.getenv(self.env_prefix + name, default)
 
     @classmethod
     def push(cls, ctx: Context) -> None: