+++ arc/types/helpers.py
@@ -14,12 +14,15 @@ from typing import (
 )
 import re
 
-from arc import utils, result
+from arc import constants, utils, result
+from arc.color import colorize, fg
+from arc.prompt import select
 
 if TYPE_CHECKING:
     from arc.typing import Annotation
     from arc.types.aliases import TypeProtocol
     from arc.context import Context
+    from arc._command.param import Param
 
 
 T = TypeVar("T")
@@ -173,3 +176,26 @@ def match(pattern: str, string: str) -> result.Result[None, str]:
         return result.Err(f"does not follow required format: {pattern}")
 
     return result.Ok()
+
+
+def input_prompt(ctx: Context, param: Param, **kwargs):
+    empty = param.default is not constants.MISSING
+
+    return (
+        ctx.prompt.input(param.get_prompt_string(), empty=empty, **kwargs)
+        or constants.MISSING
+    )
+
+
+def password_prompt(ctx: Context, param: Param):
+    return input_prompt(ctx, param, sensitive=True)
+
+
+def select_prompt(values: list, ctx: Context, param: Param, **kwargs):
+    print(param.prompt)
+    res = select(values, highlight_color=ctx.config.brand_color, **kwargs)
+
+    if res is None:
+        return constants.MISSING
+
+    return res[1]