+++ docs/docs/changelog.md
@@ -1,3 +1,12 @@
+## `6.3.0`
+
+- Added the ability to gather input from environment variables or by prompting the user directly.
+- Added `prompt.select()` for a more elegant selection menu
+
+## `6.2.2`
+
+- Fixed bug with `IoWrapper`
+
 ## `6.2.1`
 
 - Fixed command debug utilites